requirements.txt
flask==3.0.0
flask_mysqldb==2.0.0
passlib==1.7.4
wtforms==3.1.1
requests-oauthlib==1.3.1
python-dotenv==1.0.0
Flask-Login==0.6.3
